Move bootmem allocator in a sub-pkg
